requirements.txt
Flask==0.10.1
Flask-Migrate==1.4.0
Flask-SQLAlchemy==2.0
Flask-Script==2.0.5
Jinja2==2.7.3
Mako==1.0.1
MarkupSafe==0.23
PyYAML==3.11
SQLAlchemy==1.0.4
Werkzeug==0.10.4
alembic==0.7.6
blinker==1.3
cov-core==1.14.0
coverage==3.7.1
gunicorn==19.3.0
itsdangerous==0.24
mock==1.0.1
pep8==1.6.2
psycopg2==2.6
py==1.4.25
pytest==2.6.3
pytest-cov==1.8.0
requests==2.7.0
responses==0.3.0
six==1.9.0
wsgiref==0.1.2
